+++ dashi/src/lib/types/state/store.ts
@@ -1,13 +1,13 @@
+import type { StoreApi } from "zustand";
+
 import type {
   ContribPoint,
   Contributions,
   Extension,
 } from "@/lib/types/model/extension";
-import type { ApiResult } from "@/lib/utils/fetchApiResult";
 import type { ContributionState } from "@/lib/types/state/contribution";
-import type { ApiOptions } from "@/lib/api/types";
+import type { ApiOptions, ApiResult } from "@/lib/api/types";
 import type { LoggingOptions } from "@/lib/actions/helpers/configureLogging";
-import type { StoreApi } from "zustand";
 
 export interface FrameworkOptions<S extends object = object> {
   /** The host applications state management store. */
