--- dashi/src/lib/utils/isSubscriptable.ts
@@ -1,7 +1,0 @@
-export function isSubscriptable(
-  value: unknown,
-): value is { [key: string]: unknown } {
-  return (
-    (typeof value === "object" && value !== null) || typeof value === "function"
-  );
-}
